Add database connection and authentication views
- Setup and config postgresql database with psycopg2, with some utility
commands to reset the database and fill it with mock data.
- Removes unnecessary views and templates from the main __init__ file.
- Adds register/login/logout views and templates, so new users can
register via a form to create a new account with password hashing. The
login route will take email and password and check it against the hash.
- Adds a before_request hook to check the session on the request and, if
it exists, get the user from the database and store it on the g object.